app/src/main/java/com/application/newsapp/data/network/RetrofitClient.kt
package com.application.newsapp.data.network

import com.application.newsapp.utils.Constants
import retrofit2.Retrofit
import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
import retrofit2.converter.gson.GsonConverterFactory

object RetrofitClient {

    fun create(): RetrofitApi {
        val retrofit = Retrofit.Builder()
            .addConverterFactory(GsonConverterFactory.create())
            .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
            .baseUrl(Constants.API_URL)
            .build()
        return retrofit.create(RetrofitApi::class.java)
    }

}